+++ src/Symfony/Component/Console/Command/Command.php
@@ -17,6 +17,7 @@ use Symfony\Component\Console\Input\InputDefinition;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\BufferedOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Helper\HelperSet;
@@ -576,8 +577,10 @@ class Command
     public function asText()
     {
         $descriptor = new TextDescriptor();
-
-        return $descriptor->describe($this);
+        $output = new BufferedOutput(BufferedOutput::VERBOSITY_NORMAL, true);
+        $descriptor->describe($output, $this, array('raw_output' => true));
+        
+        return $output->fetch();
     }
 
     /**
@@ -592,8 +595,15 @@ class Command
     public function asXml($asDom = false)
     {
         $descriptor = new XmlDescriptor();
-
-        return $descriptor->describe($this, array('as_dom' => $asDom));
+        
+        if ($asDom) {
+            return $descriptor->getCommandDocument($this);
+        }
+        
+        $output = new BufferedOutput();
+        $descriptor->describe($output, $this);
+        
+        return $output->fetch();
     }
 
     private function validateName($name)