+++ src/Symfony/Component/HttpKernel/Kernel.php
@@ -61,11 +61,11 @@ abstract class Kernel implements KernelInterface, TerminableInterface
     protected $classes;
     protected $loadClassCache;
 
-    const VERSION         = '2.3.3-DEV';
-    const VERSION_ID      = '20303';
+    const VERSION         = '2.4.0-DEV';
+    const VERSION_ID      = '20400';
     const MAJOR_VERSION   = '2';
-    const MINOR_VERSION   = '3';
-    const RELEASE_VERSION = '3';
+    const MINOR_VERSION   = '4';
+    const RELEASE_VERSION = '0';
     const EXTRA_VERSION   = 'DEV';
 
     /**