+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+2.4.0
+-----
+
+ * added Request::getEncodings()
+ * added accessors methods to session handlers
+
 2.3.0
 -----